requirements.txt
anyjson==0.3.3
asgiref==3.2.10
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.4.0
django-json-widget==1.0.1
django-jsonfield==1.4.0
django-pgviews==0.5.5
django-query-chunk==0.0.1
django-redis==4.12.1
django-rest-swagger==2.1.2
django-silk==4.1.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-utils-six==2.0
django-utils2==2.8.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
future==0.18.2
gprof2dot==2019.11.30
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kavenegar==1.1.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
password-strength==0.0.3.post2
persiantools==1.5.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
stringcase==1.2.0
suds-py3==1.4.1.0
swagger-spec-validator==2.7.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
